getUserDefaultMt fix

+++ migrations/20240116104800_create_project_template_table.php
@@ -4,7 +4,7 @@ class CreateProjectTemplateTable extends AbstractMatecatMigration {
 
     public $sql_up = [ '
         CREATE TABLE `project_templates` (
-            `id` INT UNSIGNED NOT NULL AUTO_INCREMENT,
+            `id` bigint(20) UNSIGNED NOT NULL AUTO_INCREMENT,
             `name` VARCHAR(255) NOT NULL,
             `is_default` TINYINT(1) NOT NULL DEFAULT  0,
             `uid` bigint(20) NOT NULL,
